requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai>=1.54.0
requests==2.31.0
python-multipart==0.0.6
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-identity==1.15.0